Check ENV variables in the monitor
This allows this check to happen as an http status check and makes it
simpler since symofony will load the .env file for us.
