+++ composer.json
@@ -88,12 +88,12 @@
     "auto-scripts": {
       "cache:clear": "symfony-cmd",
       "assets:install %PUBLIC_DIR%": "symfony-cmd",
+      "monitor:health": "symfony-cmd",
       "requirements-checker": "script",
       "security-checker security:check": "script"
     },
     "ilios-scripts": [
-      "App\\Composer\\MigrateParameters::migrate",
-      "App\\Composer\\CheckRequiredENV::check"
+      "App\\Composer\\MigrateParameters::migrate"
     ],
     "post-install-cmd": [
       "@auto-scripts",
